+++ src/pages/NotFoundPage/NotFound/NotFound.module.scss
@@ -1,4 +1,4 @@
-@import "../../scss/";
+@import "../../../scss/";
 
 .imgBox {
 	display: flex;
